updated bc table
