+++ app/componentLibraryViewer/components/generalComponents.njk
@@ -17,7 +17,7 @@
     <li><a href="general/tag">Tag</a></li>
     <li><a href="general/back-to-top">Back to top</a></li>
     <li><a href="general/bc-table">BC Table</a></li>
-    <li><a href="general/bc-table-flexible">BC Table Flexible</a></li>
+    <li><a href="general/bc-table">BC Table Flexible</a></li>
     <li><a href="general/bc-error-wrapper">BC Error Wrapper</a></li>
     <li><a href="general/bc-task-list">BC Task List</a></li>
     <li><a href="general/bc-button">BC Button</a></li>
