--- app/components/general/components/bc-table-flexible/macro.njk
@@ -1,83 +1,0 @@
-{% from 'components/general/components/tag/macro.njk' import tag %}
-{% from 'components/view/components/view-data-text/macro.njk' import viewDataText %}
-{% from 'components/input/components/text-field/macro.njk' import textField %}
-{% from 'components/view/components/view-expandable-section/macro.njk' import viewExpandableSection %}
-{% from 'components/checkboxes/macro.njk' import checkboxes %}
-
-{% macro bcTableFlexible(params) %}
-  <table data-test-id="table" class="{{ param.classes }}" style="page-break-inside:auto">
-
-    {% if params.columnInfo %}
-      <thead data-test-id="table-headings" style="display:table-header-group;">
-        {% for column in params.columnInfo %}
-          <th data-test-id="column-heading-{{ loop.index0 }}" class="nhsuk-u-padding-bottom-0 {{ column.classes }}" style="width:{{ column.width }}">
-            <div data-test-id="column-heading-{{ loop.index0 }}-data" class="nhsuk-u-font-weight-bold">{{ column.data }}</div>
-            {% if column.expandableSection %}
-              <div data-test-id="column-heading-{{ loop.index0 }}-expandable" class="nhsuk-u-margin-top-2 nhsuk-u-font-weight-bold ">
-                {{ viewExpandableSection(column.expandableSection) }}
-              </div>
-            {% endif %}
-          </th>
-        {% endfor %}
-      </thead>
-    {% endif %}
-
-    <tbody>
-    {% for row in params.data %}
-      <tr data-test-id="table-row-{{ loop.index0 }}" style="page-break-inside:avoid; page-break-after:auto">
-        {% for dataPoint in row %}
-          {% set borderBottom = "none" if (dataPoint.hideSeperator) else "solid" %}
-          <td class="nhsuk-u-padding-bottom-0" style="border-bottom-style: {{ borderBottom  }}">
-
-            {% if dataPoint.href %}
-              <a data-test-id="{{ dataPoint.dataTestId }}" href="{{ dataPoint.href }}">{{ dataPoint.data }}</a>
-            {% elif dataPoint.tag %}
-              <div>
-                {{ tag({
-                  dataTestId: dataPoint.tag.dataTestId,
-                  text: dataPoint.tag.text,
-                  classes: dataPoint.tag.classes
-                }) }}
-              </div>
-            {% elif dataPoint.multiLine %}
-              <div data-test-id="{{ dataPoint.multiLine.dataTestId }}">
-                {% for line in dataPoint.multiLine.data %}
-                  <div class="dataPoint.multiLine.classes">
-                    {% if line == '' %}
-                      &nbsp;
-                    {% else %}
-                      {{ line }}
-                    {% endif %}
-                  </div>
-                {% endfor %}
-              </div>
-            {% elif dataPoint.question and dataPoint.question.type == 'input' %}
-              <div data-test-id="{{ dataPoint.dataTestId }}">
-                {{ textField(dataPoint) }}
-              </div>
-            {% elif dataPoint.question and dataPoint.question.type == 'checkbox' %}
-              <div data-test-id="{{ dataPoint.dataTestId }}">
-                {{ checkboxes({
-                  items: [dataPoint.question]
-                }) }}
-              </div>
-            {% else %}
-              {{ viewDataText({
-                dataTestId: dataPoint.dataTestId,
-                classes: dataPoint.classes,
-                data: dataPoint.data
-              }) }}
-            {% endif %}
-
-            {% if dataPoint.expandableSection %}
-              <div>
-                {{ viewExpandableSection(dataPoint.expandableSection) }}
-              </div>
-            {% endif %}
-          </td>
-        {% endfor %}
-      </tr>
-    {% endfor %}
-    </tbody>
-  </table>
-{% endmacro %}
