--- app/components/general/components/bc-table-flexible/settings.json
@@ -1,126 +1,0 @@
-{
-  "componentName": "bcTableFlexible",
-  "params": {
-    "columnInfo": [
-      {
-        "data": "Text/Link column",
-        "classes": "some-column1-class",
-        "width": "20%"
-      },
-      { 
-        "data": "Tag column",
-        "classes": "some-column2-class",
-        "width": "20%"
-      },
-      { 
-        "data": "Input column",
-        "classes": "some-column3-class",
-        "width": "40%",
-        "expandableSection": {
-          "dataTestId": "view-section-some-section-id",
-          "title": "ExpandableSection title 1",
-          "innerComponent": "Some inner text (could be html)"
-        }
-      },
-      { 
-        "data": "Checkbox column",
-        "classes": "some-column4-class",
-        "width": "20%"
-      }
-    ],
-    "data": [
-      [
-        { "data": "Some text as link", "href": "#",
-          "expandableSection": {
-            "dataTestId": "view-section-some-section-id",
-            "title": "ExpandableSection title 1",
-            "innerComponent": "Some inner text (could be html)"
-          },
-          "hideSeperator": true
-        },
-        { "tag": { "dataTestId": "a-tag-id-1", "text": "tag text", "classes": "bc-c-tag-outline" } },
-        {
-          "question": {
-            "type": "input",
-            "id": "someId",
-            "data":"The data goes here"
-          },
-          "classes": "input-classes"
-        },
-      { 
-        "question": {
-          "type": "checkbox",
-          "id": "someId1",
-          "name": "checkboxName1",
-          "value": "checkBoxValueSentInForm1",
-          "text": "some text here"
-        }
-      }
-      ],
-      [
-        { "data": "some plain text" },
-        { "tag": { "dataTestId": "a-tag-id-1", "text": "tag text", "classes": "bc-c-tag-outline" } },
-        {
-          "question": {
-            "type": "input",
-            "id": "someId",
-            "data":"The data goes here",
-            "error": { "message": "error message" }
-          },
-          "expandableSection": {
-            "dataTestId": "view-section-some-section-id",
-            "title": "ExpandableSection title 1",
-            "innerComponent": "Some inner text (could be html)"
-          },
-          "dataTestId": "input-question-id"
-        },
-        { 
-          "question": {
-            "type": "checkbox",
-            "id": "someId2",
-            "name": "checkboxName2",
-            "value": "checkBoxValueSentInForm2",
-            "text": "some other text here",
-            "checked": true
-          },
-          "dataTestId": "checkbox-question-id"
-        }
-      ],
-      [
-        { 
-          "multiLine": { 
-            "data": ["some plain text", "and another", "", "with a blank line"],
-            "dataTestId": "multiLine-id-1"
-          }
-        },
-        { "tag": { "dataTestId": "a-tag-id-1", "text": "tag text", "classes": "bc-c-tag-outline" } },
-        {
-          "question": {
-            "type": "input",
-            "id": "someId",
-            "data":"The data goes here",
-            "error": { "message": "error message" }
-          },
-          "expandableSection": {
-            "dataTestId": "view-section-some-section-id",
-            "title": "ExpandableSection title 1",
-            "innerComponent": "Some inner text (could be html)"
-          },
-          "dataTestId": "input-question-id"
-        },
-        { 
-          "question": {
-            "type": "checkbox",
-            "id": "someId2",
-            "name": "checkboxName2",
-            "value": "checkBoxValueSentInForm2",
-            "text": "some other text here",
-            "checked": true
-          },
-          "dataTestId": "checkbox-question-id"
-        }
-      ]
-    ],
-    "classes": "nhsuk-u-font-size-19"
-  }
-}
