+++ app/components/general/components/bc-table/settings.json
@@ -2,19 +2,42 @@
   "componentName": "bcTable",
   "params": {
     "columnInfo": [
-      { "data": "Text/Link column" },
-      { "data": "Tag column" },
-      { "data": "Input column",
+      {
+        "data": "Text/Link column",
+        "classes": "some-column1-class",
+        "width": "20%"
+      },
+      { 
+        "data": "Tag column",
+        "classes": "some-column2-class",
+        "width": "20%"
+      },
+      { 
+        "data": "Input column",
+        "classes": "some-column3-class",
+        "width": "40%",
         "expandableSection": {
-        "dataTestId": "view-section-some-section-id",
-        "title": "ExpandableSection title 1",
-        "innerComponent": "Some inner text (could be html)"
-      }},
-      { "data": "Checkbox column" }
+          "dataTestId": "view-section-some-section-id",
+          "title": "ExpandableSection title 1",
+          "innerComponent": "Some inner text (could be html)"
+        }
+      },
+      { 
+        "data": "Checkbox column",
+        "classes": "some-column4-class",
+        "width": "20%"
+      }
     ],
     "data": [
       [
-        { "data": "Some text as link", "href": "#" },
+        { "data": "Some text as link", "href": "#",
+          "expandableSection": {
+            "dataTestId": "view-section-some-section-id",
+            "title": "ExpandableSection title 1",
+            "innerComponent": "Some inner text (could be html)"
+          },
+          "hideSeperator": true
+        },
         { "tag": { "dataTestId": "a-tag-id-1", "text": "tag text", "classes": "bc-c-tag-outline" } },
         {
           "question": {
@@ -22,12 +45,7 @@
             "id": "someId",
             "data":"The data goes here"
           },
-          "classes": "The classes go here",
-          "expandableSection": {
-            "dataTestId": "view-section-some-section-id",
-            "title": "ExpandableSection title 1",
-            "innerComponent": "Some inner text (could be html)"
-          }
+          "classes": "input-classes"
         },
       { 
         "question": {
@@ -103,6 +121,6 @@
         }
       ]
     ],
-    "columnClass": "nhsuk-grid-column-one-quarter nhsuk-u-font-size-19"
+    "classes": "nhsuk-u-font-size-19"
   }
 }
